Updated gantry height
This is supposed to be a REAL and not a INTEGER as it is actually a double value. welp

+++ app/src/main/java/sg/edu/ntu/scse/cz2006/gymbuddies/data/CarPark.java
@@ -59,7 +59,7 @@ public class CarPark {
 
     @NonNull
     @ColumnInfo(name = "gantry_height")
-    public int gantryHeight;
+    public double gantryHeight;
 
     @NonNull
     @ColumnInfo(name = "car_park_basement")
